[bug 1195747] Handle "unknown" for country field in hb post
This is a stop gap fix to handle "unknown" as a value for the country
field for HB posts. What's going on now is that it throws an HTTP 400
error. This fixes it so that "unknown" gets changed to "UNK" which is a
valid value.

Ultimately we should fix this in HB. There's a PR to fix it there. Once
it's fixed there, we can undo this change.