[bug 1202288] Reverting to mock 1.0.1
mock 1.3.0 uses pbr to provide sane packaging stuff. However, that
doesn't seem to work in a vendor/ environment on our servers.

I didn't figure out what was causing the problem. Given that this is a
testing-related library and it doesn't really matter if it's up-to-date,
I'm just going to revert it and deal with it another day.