[bug 1203119] Update django-ratelimit from 0.4 to 0.6
This updates django-ratelimit to the latest version. There were some big
changes in django-ratelimit, so this updates the fjord code to account
for those changes.

* ratelimit and is_ratelimited function signature changes
* key function takes (group, request) as args now

I also reduced the sleep we use to make the throttle test work from 0.05
to 0.01. That works locally, but we'll have to see if that works on
travis, too.