[bug 1205013] Add comprehensive test plan
Periodically, we do substantive changes that affect the whole codebase.
We have a unit test suite and a set of smoketests, but there's a lot of
stuff that isn't covered by those.

Instead of writing up a test plan every time we do a substantive change,
this is an attempt at writing a comprehensive test plan covering all the
important things that we can reuse every time we do a substantive
change.