[bug 1209787] Fix tests sending GA data
We have a handful of google_utils tests that need to test that function.
This fixes them so they're all mocked and none of them send actual data.

We have a bunch of other tests that test code that calls google_utils
functions. This fixes the google_utils functions to not send any data
when running in a test context.