[bug 1213340] Add fields to querystring in hb data view
Gregg said it helps to be able to restrict the hb data list to just
fields he's interested in. This helps him debug his
problem-of-the-moment.

I wanted to implement this quickly, so I added a fields param to the
querystring as well as some minor stuff to make sure it's persisted via
links. I also added a list of all the fields at the bottom of the page
so Gregg can concoct his urls by hand.

If this turns out to be much used, we'll want to add a form to make it
easier to select and deselect fields.