[bug 1216624] Add --check-header to msgfmt run
This will cause msgfmt to check the headers of the .po file before it
compiles it to a .mo file. In most cases, the header values don't matter
a lot. However, this validates the Plural-Forms header which is important.

If Plural-Forms header is invalid, then msgfmt will exit with a status
code of 1 and we don't replace the existing .mo file. If it's fine,
then we copy the targettmp file to the destination and we're good to go.