[bug 1218381] Remove all the l10n completion stuff
I can't get this to work on Input plus it's fragile when running on
the servers and keeps breaking. It's way easier to put the
infrastructure elsewhere and run it there. Thus I rewrote it and run
it on people.m.o now.

Given that, I'm removing all this stuff from fjord.