[bug 1221252] Fix url field handling for feedback
This changes the url field to be 1000 characters. Further, it changes
the API and the feedback form handler to allow urls bigger than 1000
characters, but truncate them. It adds a test for the feedback form
handler and fixes the api test. On top of that, it centralizes the
magic number to the config file so I can more easily change it.