[bug 1221352] Handle emoji
This adds a function that either removes or converts emoji to names
depending on whether unicodedata2 is installed and whether the name
exists.

unicodedata2 is the Python2 backport for unicodedata updates in
Python3. It's a binary dependency, so it needs to be installed on all
the servers directly. That requires webops to make changes to the
servers which they can't do now because we're in a server freeze.

Because of that, I wrote the code such that if unicodedata2 isn't
installed, it'll remove the emoji. I'll write up a bug to install
unicodedata2 on all the servers. Once that happens, we can change the
code so that we expect unicodedata2 to be installed.