[bug 1223625] Add source and api params to secret histogram api
I also redid the api code in ES because previously if a feedback
wasn't saved with the API, then the value would be None which ES
doesn't index and it's harder to query. This changes it so that the
api values are all strings and the None value is now "None".