[bug 1229366] Pin hb data view and improve logging
Previously, I think it was possible for Input to stomp on a new data
packet with an old one because the updated_ts check would be using on
stale data because we're in a multidb environment and there's
replication lag and all that.

This pins that view so that it always happens on the master db and thus
always sees the current view of the world.

Further, this adds a bunch of statsd calls so we can measure various
scenarios we don't otherwise have a window into.