[bug 1235423] Refactor feedback_router
This splits the picker view from the feedback_router code. Having said
that, the oldandroid code makes that continue to look a bit goofy.

This drops the requirement that we *only* show the Firefox OS form to
users of Firefox OS. With these code changes, if they got to
/feedback/, they'll now go to the picker. We don't get a lot of
Firefox OS feedback. As of 1.3, they have an in-product form that uses
the API anyhow.

This gets rid of the PRODUCT_OVERRIDES scaffolding which we haven't
used in ages and has very much overstayed its welcome.

The end result is a feedback_router that's a little more
straight-forward.