[bug 1235423] Refactor around thank you view
This reworks the feedback flow a bit more so that adding a new form
entails writing a function with a decorator that registers it for a
specific product.

    @register_product_config(product_slug, thanks_template)

It's still a little goofy in that this doesn't read quite right, but
it's good enough for now especially where we only have 2 forms
currently but will have 3 in the near future. We can flesh things out
and make it more featureful when needed.