[bug 1215125] Fix whimsical locale in admin
Django 1.8 changed trans_real.activate such that if you pass in a ''
language, then it does nothing. Since prefixer.locale is '' when
handling an /admin/ request, the request gets handled with the
previously handled request's activated translations.

This fixes that so that if prefixer.locale is '', then we use the
settings.LANGUAGE_CODE to activate the translations.