Fix bug 1264916: Always pass a str to parse_qsl.
When a URL contains unicode, Django's request.get_full_path() returns
a unicode string with percent-encoded unicode strings. parse_qsl fails
to parse this properly, and returns a unicode string with bytes in it,
instead of codepoints. To fix this, we encode the URL in ASCII before
passing it to parse_qsl.